+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/metric/makers/AggregationAverageMaker.java
@@ -9,6 +9,8 @@ import com.yahoo.bard.webservice.data.metric.LogicalMetric;
 import com.yahoo.bard.webservice.data.metric.LogicalMetricInfo;
 import com.yahoo.bard.webservice.data.metric.MetricDictionary;
 import com.yahoo.bard.webservice.data.metric.TemplateDruidQuery;
+import com.yahoo.bard.webservice.data.metric.mappers.NoOpResultSetMapper;
+import com.yahoo.bard.webservice.data.metric.mappers.ResultSetMapper;
 import com.yahoo.bard.webservice.data.metric.protocol.DefaultSystemMetricProtocols;
 import com.yahoo.bard.webservice.data.metric.protocol.ProtocolSupport;
 import com.yahoo.bard.webservice.data.metric.protocol.protocols.ReaggregationProtocol;
@@ -24,6 +26,7 @@ import com.yahoo.bard.webservice.druid.model.postaggregation.FieldAccessorPostAg
 import com.yahoo.bard.webservice.druid.model.postaggregation.PostAggregation;
 
 import com.google.common.collect.ImmutableSet;
+import com.yahoo.bard.webservice.web.ChainingResultSetMapper;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -231,5 +234,19 @@ public class AggregationAverageMaker extends BaseProtocolMetricMaker {
         return originalSourceMetric instanceof SketchAggregation ?
                 getSketchConverter().asSketchEstimate((SketchAggregation) originalSourceMetric) :
                 originalSourceMetric;
+    }
+
+    @Override
+    protected ResultSetMapper makeCalculation(
+            final LogicalMetricInfo logicalMetricInfo, final List<LogicalMetric> dependentMetrics
+    ) {
+        if (!(dependentMetrics.get(0).getCalculation() instanceof NoOpResultSetMapper)) {
+            return ChainingResultSetMapper.createAndRenameResultSetMapperChain(
+                    logicalMetricInfo,
+                    dependentMetrics.get(0).getCalculation()
+            );
+        }
+
+        return NO_OP_MAPPER;
     }
 }